+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+- Fix support for connection Upgrade and CONNECT when some data in the stream has been read. (#882)
+
 ## 1.0.3 (February 13th, 2024)
 
 - Fix support for async cancellations. (#880)